# docs/swagger.yaml
basePath: /api
definitions:
  cmd.AnalyticsEvent:
    properties:
      boothID:
        type: string
      createdAt:
        type: string
      eventName:
        type: string
      id:
        type: string
      payload:
        additionalProperties: {}
        type: object
      sessionID:
        type: string
    type: object
  cmd.Booth:
    properties:
      branchID:
        type: string
      config:
        additionalProperties: {}
        type: object
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/go-ddd-clean_internal_domain_booth.BoothStatus'
      tokenVersion:
        type: integer
      type:
        $ref: '#/definitions/go-ddd-clean_internal_domain_booth.BoothType'
    type: object
  cmd.BoothAnalyticsEventRequest:
    properties:
      event_name:
        type: string
      payload:
        additionalProperties: {}
        type: object
      session_id:
        type: string
    type: object
  cmd.BoothCreateRequest:
    properties:
      branch_id:
        type: string
      config:
        additionalProperties: {}
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  cmd.BoothLog:
    properties:
      boothID:
        type: string
      createdAt:
        type: string
      eventType:
        type: string
      id:
        type: string
      level:
        $ref: '#/definitions/go-ddd-clean_internal_domain_logging.Level'
      message:
        type: string
    type: object
  cmd.BoothLogCreateRequest:
    properties:
      event_type:
        type: string
      level:
        type: string
      message:
        type: string
    type: object
...
  cmd.BoothUpdateRequest:
    properties:
      branch_id:
        type: string
      config:
        additionalProperties: {}
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  cmd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  go-ddd-clean_internal_domain_booth.BoothStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - BoothStatusActive
    - BoothStatusInactive
  go-ddd-clean_internal_domain_booth.BoothType:
    enum:
    - physical
    - virtual
    type: string
    x-enum-varnames:
    - BoothTypePhysical
    - BoothTypeVirtual
  go-ddd-clean_internal_domain_logging.Level:
    enum:
    - info
    - warn
    - error
    type: string
    x-enum-varnames:
    - LevelInfo
    - LevelWarn
    - LevelError
...
info:
  contact: {}
  description: API documentation for the go-ddd-clean service.
  title: Photobooth Platforms API
  version: "1.0"
paths:
...
  /api/booths:
    get:
      parameters:
      - description: กรองตามสาขา
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Booth'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงรายการบูธ
      tags:
      - Booths
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลบูธ
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.BoothCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Booth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างบูธใหม่
      tags:
      - Booths
  /api/booths/{id}:
    delete:
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ลบบูธ
      tags:
      - Booths
    get:
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Booth'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลบูธ
      tags:
      - Booths
    put:
      consumes:
      - application/json
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่ต้องการแก้ไข
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.BoothUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Booth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ปรับปรุงข้อมูลบูธ
      tags:
      - Booths
  /api/booths/{id}/analytics:
    get:
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      - description: จำนวนรายการ (ค่าเริ่มต้น 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.AnalyticsEvent'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงข้อมูลการใช้งานบูธ
      tags:
      - Booth Analytics
    post:
      consumes:
      - application/json
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลเหตุการณ์
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.BoothAnalyticsEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.AnalyticsEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างข้อมูลการใช้งานบูธ
      tags:
      - Booth Analytics
  /api/booths/{id}/logs:
    get:
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      - description: จำนวนรายการ (ค่าเริ่มต้น 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.BoothLog'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงบันทึกของบูธ
      tags:
      - Booth Logs
    post:
      consumes:
      - application/json
      parameters:
      - description: รหัสบูธ
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลบันทึก
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.BoothLogCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.BoothLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: บันทึกเหตุการณ์ของบูธ
      tags:
      - Booth Logs
...
schemes:
- http
securityDefinitions:
  BoothTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"